Allow the status change to be turned off
Some of our code bases are old and un-linted. Sometimes we ignore the linting
issue as it gets in the way of us doing actual changes, but it's still nice to
see the comments on a review.

If we were to use the latest version of lint-review, then we can't use the
branch protection feature from GitHub
(https://github.com/blog/2051-protected-branches-and-required-status-checks)