Update tests to support newer versions of ESLint.
It looks like the original tests were written for a pre 1.0 release of
ESLint.  The latest version is now 2.4.0.  The breaking changes are
described here:
http://eslint.org/docs/user-guide/migrating-to-1.0.0
http://eslint.org/docs/user-guide/migrating-to-2.0.0

The biggest change is that all rules are off by default.  This means
that tests depending on the default ruleset don't work.  That original
behavior can be approximated by making a config which extends
"eslint:recommended".