Remove git API based approach.
While it solved the authorship issues for the commit strategy, it was
incompatible with forks. We cannot safely assume that our bot account
will have access to the fork. Because of this we will be unable to
create the necessary tree/commit/ref update. However, git-cli works
because of the special permissions github affords pushes to branches
involved in pull requests.

There is still one gap where if a contributor disables maintainer
updates, we'll hard fail but I'll fix that soon.