Update tests, fixtures for github3.py 1.2.0
- Update initialization for various github3.py objects to include
a GitHubSession instance

- tests/__init__.py: Import ShortCommit instead of RepoCommit
The API for RepoCommit appears to be different in the newer version of
github3. Specifically, RepoCommit doesn't have an `author` attribute

- tests/fixtures/commits.json: Updated to work with ShortCommit
- tests/fixtures/repository.json: Dumped markstory/lint-test to json
The newer Repository object expects fields that were missing from the
old fixture

- tests/fixtures/pull_request.json: Dumped markstory/lint-test/pulls/1 to json
This is just dumped to json using the newer github3 library

PR #1 is different in part because the current HEAD isn't at
https://github.com/markstory/lint-test/pull/1/commits/53cb70abadcb3237dcb2aa2b1f24dcf7bcc7d68e
anymore.

I also may have changed "markstory" to "contributor" in more places in
the PR `head` section than it was in the old fixture.

- Updated PullRequest object creation from pull_request.json throughout
It looks like the old fixture was possibly a hook payload, thus the pull
request body was accessed via the `pull_request` key from the fixture,
whereas the new fixture is a serialized PullRequest object

- tests/test_repo.py: Updated TestGithubRepository.test_head to reflect updated commit for
markstory/lint-test/pull/1

- tests/fixtures/comments_current.json; tests/fixtures/comments_none_current.json
Add additional fixture fields required by newer IssueComment